requirements.txt
beautifulsoup4==4.10.0
boto3==1.21.32
botocore==1.24.32
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
MarkupSafe==2.1.1
python-dateutil==2.8.2
python-decouple==3.6
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
soupsieve==2.3.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
